Fixes Page Content Overflowing on Blog on Small Displays
I tested https://www.almalinux.com/blog on an iPhone 12 pro, and a Samsung S20 and the content is wider than the screen.

The cause was the inline CSS in the /layouts/blog/list.html

I moved the CSS into the CSS files, and changed

width: 450px;

to

max-width: 100%;

This will prevent page overflow on devices with > 800px wide displays.