Tightens Up Code Blocks
This tightens up code blocks as well as removes some unneeded CSS.

I know font sizes were mentioned as an issue.  Currently we use a monospace font for code blocks, and a regular font for all other content.

We could change code blocks to use the same font that all the other content uses if we want to be uniform.