Adds API Search
Their default search using their UI script got us VERY CLOSE to what we needed but unfortunately it wasn't close enough.

There is a lot of logic spent on maintaining currently checked checkboxes when a new search is done (each time you click a checkbox a new search is performed).

It also maintains the OPEN status of the details element.

This needs work to make it more mobile friendly.

This needs work to account for pagination.

This needs work to show actual search results but I am going to need to get some feedback on what exactly  we need for that.