FIPS 140 FAQ fixes
Fixed "processes" typo.

Fixed use of "compliant" term - should be "validated".

Made version clearer - we are 140-3 validated as 140-2 will be deprecated next year.

Updated progress - we are now on the Active list, not in the process.

Pointed "read more" link to the newer blog post by @Noam-Alum (which links back to my older one too).